fix: 上場前掃描抓到 5 個靜默失敗的 blocker
明天要驗的路徑逐條審過（4 個 agent 平行掃描），全部是「不報錯但不動作」
那一類——現場最吃時間的。

## ★★ 最嚴重：VIP 迎賓與黑名單通報在程式上不可能觸發

    user_auth_node._publish_identity 從未設 msg.similarity   -> 恆為 0.0
    bank_reception_node:146  VIP/BLACKLIST 有一道
        similarity < min_confidence(0.5) 就 return
    -> 0.0 < 0.5 恆成立 -> 兩個劇本永遠不會跑

節點在跑、/user_identity 有發、recognized 也是 True，只有「劇本不動」，
log 一行都沒有。而這正是專題的主秀，也是明天第一優先要驗的。

_process_face_recognition 明明算出了 max_similarity 卻在 return 時丟掉。
改成回傳 (uuid, similarity)，未認出時也回實際最高分——E2 誤認率分析要的
就是「未註冊者拿到多少分」的分布。visit_log.db 的 confidence 欄一併修好。

## 人臉線另外兩個靜默失敗

1. face_embedding_node 用 CompressedImage 訂 /camera/color/image_raw，
   但相機在那條話題發的是 sensor_msgs/Image（astra_camera 直接
   create_publisher<Image>，沒走 image_transport）。型別對不上 ->
   訂閱建立成功、log 照印「訂閱話題: ...」、一則影像都收不到。
   而 user_auth 的預設值本來就是 .../compressed —— 兩個節點連話題都不同。
   預設值改成 compressed 那條。

2. face_engine.py:120 bbox 用 .astype(int)，但 FaceEmbedding.msg 是
   float32[4]。rosidl 的 setter 會 assert isinstance(v, float)，Python int
   不是 float 的 instance -> 每偵測到一張臉就 AssertionError，被上層
   except 吃掉。同一個寫法也讓 ExtractFaceEmbedding 永遠回 success=False
   （HMI 照片註冊一起壞）。改成 float32 並在指派處明確轉 float。

## 量測工具

steer_asym_check_cc.py 發到 cmd_vel_smoothed，但那條話題唯一的訂閱者是
steering_trim_cc，而它只在 nav_bringup 裡啟動——本檔用法卻寫「只跑
run_sensors_cc.sh」。四段全部不會動，還是照印完整結果表（每段走 0.000 m），
三個判定區塊因 arc>0.05 沒過而靜默跳過。
而且這支本來就不該經過 trim：它量的是硬體原始不對稱。改直接發 cmd_vel
（對照 odom_gyro_check_cc.py:101 同樣做法）。

## HMI 朗讀兩個只在平板重現的坑

1. tts.on 從 localStorage 還原成 true 時，iOS/Safari 的解鎖沒發生過
   （那把鑰匙只在 onclick 從關轉開時才轉，而按鈕已經是開的，
   操作者不會去按）。整場沒聲音、無錯誤訊息，桌機 Chrome 重現不出來。
   改成掛一次性 pointerdown 解鎖，解鎖前按鈕明講「點一下啟用朗讀」。

2. ttsOnChat 在對話為空時提前 return，seeded 沒設起來，於是第一句真正的
   回覆被當成歷史訊息吞掉——機器人的第一句永遠不會被唸。
   而明天的動線正好是「開頁 -> 按朗讀 -> 打一句 -> 平板安靜」。

+++ welcoming_robot_ws/src/smartnav_ws/src/smartnav_vision/smartnav_vision/face_engine.py
@@ -117,7 +117,15 @@ class FaceEngine:
                 return None
 
             face = FaceDetectResult(
-                bbox=faces[0].bbox.astype(int),
+                # ★ 2026-08-10：原本是 .astype(int)，但 FaceEmbedding.msg 的
+                #   欄位是 `float32[4] bbox`。rosidl 產生的 setter 會 assert
+                #   每個值 isinstance(v, float)，而 Python int 不是 float 的
+                #   instance -> 每偵測到一張臉就 AssertionError，被上層的
+                #   `except Exception` 吃掉，只在 log 印一行「處理影像時發生錯誤」，
+                #   節點照跑、話題存在、但 /face_embedding 一則都發不出去。
+                #   同一個寫法也讓 ExtractFaceEmbedding 服務永遠回 success=False
+                #   （HMI 的照片註冊 /register_face_photo 一起壞）。
+                bbox=faces[0].bbox.astype(np.float32),
                 embedding=faces[0].embedding.astype(np.float32),
             )
